+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/__init__.py
@@ -13,7 +13,7 @@ from ._api_version import DocumentTranslationApiVersion
 from ._polling import DocumentTranslationPoller
 from ._models import (
     TranslationTarget,
-    JobStatusResult,
+    TranslationStatusResult,
     DocumentStatusResult,
     DocumentTranslationError,
     TranslationGlossary,
@@ -32,7 +32,7 @@ __all__ = [
     "StorageInputType",
     "FileFormat",
     "TranslationTarget",
-    "JobStatusResult",
+    "TranslationStatusResult",
     "DocumentStatusResult",
     "DocumentTranslationError",
     "DocumentTranslationPoller"