+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_polling.py
@@ -18,7 +18,7 @@ from azure.core.polling.base_polling import (
 
 from azure.core.exceptions import HttpResponseError, ODataV4Format
 from ._generated.models import TranslationStatus
-from ._models import JobStatusResult
+from ._models import TranslationStatusResult
 
 
 if TYPE_CHECKING:
@@ -54,12 +54,12 @@ class DocumentTranslationPoller(LROPoller):
 
     @property
     def details(self):
-        # type: () -> JobStatusResult
+        # type: () -> TranslationStatusResult
         """The details for the translation operation
 
-        :return: JobStatusResult
+        :return: ~azure.ai.translation.document.TranslationStatusResult
         """
-        return JobStatusResult._from_generated(self._polling_method._current_body)  # pylint: disable=protected-access
+        return TranslationStatusResult._from_generated(self._polling_method._current_body)  # pylint: disable=protected-access
 
     @classmethod
     def from_continuation_token(cls, polling_method, continuation_token, **kwargs):