+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/aio/_async_polling.py
@@ -12,7 +12,7 @@ from azure.core.polling.base_polling import (
 )
 from azure.core.polling.async_base_polling import AsyncLROBasePolling
 from .._generated.models import TranslationStatus
-from .._models import JobStatusResult
+from .._models import TranslationStatusResult
 
 PollingReturnType = TypeVar("PollingReturnType")
 _FINISHED = frozenset(["succeeded", "cancelled", "cancelling", "failed"])
@@ -34,12 +34,12 @@ class AsyncDocumentTranslationPoller(AsyncLROPoller[PollingReturnType]):
         return self._polling_method._get_id_from_headers()  # pylint: disable=protected-access
 
     @property
-    def details(self) -> JobStatusResult:
+    def details(self) -> TranslationStatusResult:
         """The details for the translation operation
 
-        :return: JobStatusResult
+        :return: ~azure.ai.translation.document.TranslationStatusResult
         """
-        return JobStatusResult._from_generated(self._polling_method._current_body)  # pylint: disable=protected-access
+        return TranslationStatusResult._from_generated(self._polling_method._current_body)  # pylint: disable=protected-access
 
     @classmethod
     def from_continuation_token(